requirements.txt
beautifulsoup4==4.3.2
distribute==0.6.24
python-dateutil==2.2
requests==2.3.0
six==1.7.3
